requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
pandas>=1.3.0
opencv-python>=4.5.3
scipy>=1.7.1
scikit-image>=0.18.3
scikit-learn>=0.24.2
faiss-cpu>=1.7.2
matplotlib>=3.4.2
tqdm>=4.62.0
h5py>=3.3.0
